+++ python/sdss_access/sync/__init__.py
@@ -2,4 +2,7 @@ from .auth import Auth
 from .cli import Cli
 from .stream import Stream
 from .http import HttpAccess
+from .baseaccess import BaseAccess
 from .rsync import RsyncAccess
+from .curl import CurlAccess
+from .access import Access